-remove compiler warnings when building 64-bit

+++ contrib/ares/ares_process.c
@@ -587,9 +587,9 @@ static int open_tcp_socket(ares_channel channel, struct server_state *server)
   /* Acquire a socket. */
 #ifdef USE_IPV6
   assert(server->family == AF_INET || server->family == AF_INET6);
-  s = socket(server->family, SOCK_STREAM, 0);
+  s = (int)socket(server->family, SOCK_STREAM, 0);
 #else
-  s = socket(AF_INET, SOCK_STREAM, 0);
+  s = (int)socket(AF_INET, SOCK_STREAM, 0);
 #endif
   if (s == -1)
     return -1;
@@ -669,7 +669,7 @@ static int open_udp_socket(ares_channel channel, struct server_state *server)
 
   /* Acquire a socket. */
   assert(server->family == AF_INET || server->family == AF_INET6);
-  s = socket(server->family, SOCK_DGRAM, 0);
+  s = (int)socket(server->family, SOCK_DGRAM, 0);
 
   if (s == -1)
     return -1;
